Started working on serialization

+++ src/main/java/it/polimi/ingsw/model/Char1.java
@@ -15,7 +15,7 @@ public class Char1 implements CharacterCard {
     }
 
     @Override
-    public int getId() {
+    public byte getId() {
         return 1;
     }
