+++ src/main/java/it/polimi/ingsw/model/Char2.java
@@ -19,7 +19,7 @@ public class Char2 implements CharacterCard {
     }
 
     @Override
-    public int getId() {
+    public byte getId() {
         return 2;
     }
