+++ src/main/java/it/polimi/ingsw/model/CharacterCard.java
@@ -6,6 +6,6 @@ import it.polimi.ingsw.exceptions.GameException;
 public interface CharacterCard {
     void play(ExpertGame game) throws GameException, EndGameException;
     byte getCost();
-    int getId();
+    byte getId();
     boolean canPlay(int nInput);
 }
